fix: make IFD0 offset guard overflow-safe (Long arithmetic)
Consistency with the entry-level Long guards; the prior Int `ifd0 + 2`
could overflow for a crafted near-MAX offset (non-exploitable — the outer
runCatching caught it — but the guard is now self-sufficient).

+++ core/src/main/kotlin/com/forensics/core/handler/exif/JpegExifHandler.kt
@@ -46,7 +46,7 @@ class JpegExifHandler : FormatHandler {
         val littleEndian = String(bytes, tiffBase, 2, Charsets.US_ASCII) == "II"
         val tiff = TiffReader(bytes.copyOfRange(tiffBase, bytes.size), littleEndian)
         val ifd0 = tiff.u32(4).toInt()
-        if (ifd0 < 0 || ifd0 + 2 > tiff.data.size) return emptyList()
+        if (ifd0 < 0 || ifd0.toLong() + 2 > tiff.data.size) return emptyList()
 
         val count = tiff.u16(ifd0)
         val fields = ArrayList<MetadataField>()
